--- ogma/.~lock.defs.csv#
@@ -1,1 +1,0 @@
-,kurt,kurt-laptop,20.01.2022 16:15,file:///home/kurt/.config/libreoffice/4;
