+++ ogma/tests/commands/mod.rs
@@ -1,4 +1,10 @@
 
 
 mod arithmetic;
+mod cmp;
 mod diagnostics;
+mod io;
+mod logic;
+mod morphism;
+mod types;
+mod pipeline;
